+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 0.1.3
+- Add verification and UX around `rubyfmt` availability before running the formatter.
+
 ## 0.1.2
 - Disable **Format Selection** by default since `rubyfmt` does not support range formatting.
   - Add `formatto.enableRangeFormatting` setting to opt in to an experimental heuristic.